// types.ts
export enum UserRole {
  PROCESSING_WORKER = 'Processing Worker',
  PACKING_STAFF = 'Packing Staff',
  FINANCE_CLERK = 'Finance Clerk',
  ADMIN = 'Admin'
}